Extract setupGlobals to shared test helpers
Required to write tests for models using
`entryTypeEditorControllerUrls` outside of `pageflow` package.

REDMINE-17429
