+++ package/documentation.yml
@@ -126,4 +126,5 @@ toc:
       Utility functions exported by `pageflow/testHelpers`.
     children:
       - factories
+      - setupGlobals
       - useFakeTranslations
