+++ package/spec/support/index.js
@@ -1,7 +1,6 @@
 export * from 'pageflow/testHelpers';
 
 export * from './fakeEventEmitter';
-export * from './setupGlobals';
 
 export * from './useFakeXhr';
 export * from './useHtmlSandbox'
