+++ package/src/testHelpers/index.js
@@ -1,4 +1,5 @@
 export * from './dominos/editor';
 export * from './dominos/ui';
 export * from './factories';
+export * from './setupGlobals';
 export * from './useFakeTranslations';
